Assignment 1 Problem 1
Write a method duplicate Stack that returns a new stack containing the same
elements and in the same order as the stack specified in the parameter. The
method should create a new stack and fill it with the same data elements as the
given stack. (You do not need to duplicate the contents of the elements.).
Before the method finishes, it must restore the contents of the original stack to
its original state (same contents in the same order).